update post and user seed to get more data and add validation for only allow seeding on development

+++ src/seed/seed.service.ts
@@ -1,19 +1,25 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable, NotAcceptableException } from '@nestjs/common'
 import { User } from 'src/auth/user.entity'
 import { UserRepository } from 'src/auth/user.repository'
 import { Post } from 'src/post/entities/post.entity'
 import { PostRepository } from '../post/post.repository'
 import { UserSeed, SeedPosts } from './data'
 import * as bcrypt from 'bcrypt'
+import { ConfigService } from '@nestjs/config'
 
 @Injectable()
 export class SeedService {
   constructor(
     private readonly userRepository: UserRepository,
-    private readonly postRepository: PostRepository
+    private readonly postRepository: PostRepository,
+    private readonly configService: ConfigService
   ) {}
 
   async executeAllSeeds() {
+    if (!(this.configService.get('ENVIROMENT') === 'DEV')) {
+      throw new NotAcceptableException()
+    }
+
     await this.deleteAllTables()
     const users: User[] = []
     const posts: Post[] = []
@@ -31,6 +37,12 @@ export class SeedService {
     for (const p of SeedPosts) {
       const post = this.postRepository.create(p)
       post.user = users[0]
+      await this.postRepository.save(post)
+      posts.push(post)
+    }
+    for (const p of SeedPosts) {
+      const post = this.postRepository.create(p)
+      post.user = users[1]
       await this.postRepository.save(post)
       posts.push(post)
     }
